Directions refactoring
