+++ app/src/main/res/layout/settings_bottom_sheet.xml
@@ -4,19 +4,19 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/settings_bottom_sheet"
     android:layout_width="match_parent"
-    android:layout_height="100dp"
+    android:layout_height="@dimen/bottom_sheet_height"
     android:orientation="vertical"
     app:behavior_hideable="false"
-    app:behavior_peekHeight="20dp"
+    app:behavior_peekHeight="@dimen/peek_height"
     app:layout_behavior="android.support.design.widget.BottomSheetBehavior"
     >
 
   <TextView
       android:layout_width="match_parent"
-      android:layout_height="20dp"
+      android:layout_height="@dimen/peek_height"
       android:background="@color/colorAccent"
       android:gravity="center"
-      android:text="MODE"
+      android:text="@string/bottom_sheet_peek_name"
       android:textColor="@android:color/white" />
 
   <LinearLayout
@@ -28,7 +28,7 @@
 
     <Switch
         android:id="@+id/night_mode_sw"
-        android:text="Night mode "
+        android:text="@string/night_mode"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         />
