+++ app/src/main/res/values/dimens.xml
@@ -1,3 +1,5 @@
 <resources>
   <dimen name="fab_margin">16dp</dimen>
+  <dimen name="peek_height">28dp</dimen>
+  <dimen name="bottom_sheet_height">100dp</dimen>
 </resources>
