+++ app/src/main/res/values/strings.xml
@@ -5,4 +5,6 @@
   <string name="black_theme">black white theme</string>
   <string name="settings">Settings</string>
   <string name="theme_changed">The theme has changed</string>
+  <string name="bottom_sheet_peek_name">MODE</string>
+  <string name="night_mode">Night mode</string>
 </resources>
